move grpc client to cmd
