--- types/jsonrpc/types.go
@@ -1,1 +1,0 @@
-package jsonrpc
